Created a new RabbitMQ config file to centralize the creation of ConnectionFactory and the message serialization/deserialization implementation
