+++ backend/src/main/java/com/audiosource/backend/messaging/config/AudioRabbitMQConfig.java
@@ -4,31 +4,12 @@ import org.springframework.amqp.core.Binding;
 import org.springframework.amqp.core.BindingBuilder;
 import org.springframework.amqp.core.DirectExchange;
 import org.springframework.amqp.core.Queue;
-import org.springframework.amqp.rabbit.config.RetryInterceptorBuilder;
-import org.springframework.amqp.rabbit.config.SimpleRabbitListenerContainerFactory;
-import org.springframework.amqp.rabbit.connection.CachingConnectionFactory;
-import org.springframework.amqp.rabbit.connection.ConnectionFactory;
-import org.springframework.amqp.rabbit.core.RabbitTemplate;
-import org.springframework.amqp.rabbit.listener.RabbitListenerContainerFactory;
-import org.springframework.amqp.support.converter.Jackson2JsonMessageConverter;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 public class AudioRabbitMQConfig {
-
-    @Value("${spring.rabbitmq.host}")
-    private String rabbitmqHost;
-
-    @Value("${spring.rabbitmq.port}")
-    private int rabbitmqPort;
-
-    @Value("${spring.rabbitmq.username}")
-    private String rabbitmqUsername;
-
-    @Value("${spring.rabbitmq.password}")
-    private String rabbitmqPassword;
 
     @Value("${audioFiles.queue.name}")
     private String audioFilesQueueName;
@@ -55,42 +36,5 @@ public class AudioRabbitMQConfig {
     @Bean
     public Binding audioFilesBinding(Queue audioFilesQueue, DirectExchange audioExchange) {
         return BindingBuilder.bind(audioFilesQueue).to(audioExchange).with(audioFilesRoutingKey);
-    }
-
-    /* ConnectionFactory bean to establish a connection to RabbitMQ */
-    @Bean
-    public ConnectionFactory connectionFactory() {
-        CachingConnectionFactory factory = new CachingConnectionFactory();
-        factory.setHost(rabbitmqHost);
-        factory.setPort(rabbitmqPort);
-        factory.setUsername(rabbitmqUsername);
-        factory.setPassword(rabbitmqPassword);
-        return factory;
-    }
-
-    /* Message Serialization for producers : Ensures that outgoing messages are automatically converted
-    from Java objects to JSON when sent to RabbitMQ.*/
-    @Bean
-    public RabbitTemplate rabbitTemplate(ConnectionFactory connectionFactory) {
-        RabbitTemplate template = new RabbitTemplate(connectionFactory);
-        template.setMessageConverter(new Jackson2JsonMessageConverter());
-        return template;
-    }
-
-    /* Message Deserialization for consumers : Ensures that the incoming messages are converted back
-    * from JSON to the corresponding Java objects. */
-    @Bean
-    public RabbitListenerContainerFactory<?> rabbitListenerContainerFactory(ConnectionFactory connectionFactory) {
-        SimpleRabbitListenerContainerFactory factory = new SimpleRabbitListenerContainerFactory();
-        factory.setConnectionFactory(connectionFactory);
-        factory.setMessageConverter(new Jackson2JsonMessageConverter());
-
-        // Retry policy for message handling
-        factory.setAdviceChain(RetryInterceptorBuilder
-                .stateless()
-                .maxAttempts(5)
-                .backOffOptions(1000, 2.0, 10000) // Initial interval, multiplier, max interval
-                .build());
-        return factory;
     }
 }
