+++ backend/src/main/resources/application.properties
@@ -38,7 +38,7 @@ audioFiles.routing.key=audio.routing.key
 # ProcessedFilesQueue : Name, exchange and routing key
 processedFiles.queue.name=processedFilesQueue
 processedFiles.exchange.name=processingExchange
-procesedFiles.routing.key=processed.routing.key
+processedFiles.routing.key=processed.routing.key
 
 # ErrorQueue : Name, exchange and routing key
 error.queue.name=errorQueue
